fix: Fix imports after moving the "Dashboard" Thumb components to the DashboardThumb folder
